+++ src/main/java/cn/haitaoss/tinyspringmvc/framework/handlerAdapter/ControllerHandlerAdapter.java
@@ -20,9 +20,7 @@ public class ControllerHandlerAdapter extends AbstractHandlerAdapter {
 
     @Override
     public boolean supports(Object handler) {
-        if (handler instanceof Controller)
-            return true;
-        return false;
+        return handler instanceof Controller;
     }
 
     @Override